watch & reload files when running workers
