+++ src/server/workers/cycleLaunched/index.js
@@ -1,1 +1,4 @@
 require('./worker').start()
+
+// file watch & reload
+require('src/server/configureWatcher')()
